+++ src/container/SequentialContainer/Deque.ts
@@ -73,7 +73,7 @@ export class DequeIterator<T> extends ContainerIterator<T> {
  * reason: 当不停向后或向前 push 时一旦指针到达末尾就会扩容，即使空间足够
  */
 class Deque<T> extends SequentialContainer<T> {
-  private static sigma = 3;
+  private static sigma = 2;
   private static initBucketSize = (1 << 12);
   private first = 0;
   private curFirst = 0;
@@ -103,9 +103,6 @@ class Deque<T> extends SequentialContainer<T> {
     for (let i = 0; i < this.bucketNum; ++i) {
       this.map.push(new Array(this.bucketSize));
     }
-    const needBucketNum = Math.ceil(_length / this.bucketSize);
-    this.first = Math.floor(this.bucketNum / 2) - Math.floor(needBucketNum / 2);
-    this.last = this.first;
     container.forEach(element => this.pushBack(element));
     this.size = this.size.bind(this);
     this.getElementByPos = this.getElementByPos.bind(this);
@@ -119,15 +116,15 @@ class Deque<T> extends SequentialContainer<T> {
       newMap.push(new Array(this.bucketSize));
     }
     const needBucketNum = Math.ceil(originalSize / this.bucketSize);
-    const newFirst = Math.floor(newBucketNum / 2) - Math.floor(needBucketNum / 2);
-    let newLast = newFirst; let newCurLast = 0;
+    let newLast = 0;
+    let newCurLast = 0;
     if (this.length) {
       for (let i = 0; i < needBucketNum; ++i) {
         for (let j = 0; j < this.bucketSize; ++j) {
-          newMap[newFirst + i][j] = this.front();
+          newMap[i][j] = this.front();
           this.popFront();
           if (!this.length) {
-            newLast = newFirst + i;
+            newLast = i;
             newCurLast = j;
             break;
           }
@@ -136,7 +133,7 @@ class Deque<T> extends SequentialContainer<T> {
       }
     }
     this.map = newMap;
-    this.first = newFirst;
+    this.first = 0;
     this.curFirst = 0;
     this.last = newLast;
     this.curLast = newCurLast;
@@ -144,9 +141,9 @@ class Deque<T> extends SequentialContainer<T> {
     this.length = originalSize;
   }
   private getElementIndex(pos: number) {
+    let curNodeBucketIndex, curNodePointerIndex;
     const curFirstIndex = this.first * this.bucketSize + this.curFirst;
     const curNodeIndex = curFirstIndex + pos;
-    let curNodeBucketIndex, curNodePointerIndex;
     if ((curNodeIndex + 1) % this.bucketSize === 0) {
       curNodeBucketIndex = (curNodeIndex + 1) / this.bucketSize - 1;
       curNodePointerIndex = this.bucketSize - 1;
@@ -154,18 +151,8 @@ class Deque<T> extends SequentialContainer<T> {
       curNodeBucketIndex = Math.floor((curNodeIndex + 1) / this.bucketSize);
       curNodePointerIndex = (curNodeIndex + 1) % this.bucketSize - 1;
     }
+    curNodeBucketIndex %= this.bucketNum;
     return { curNodeBucketIndex, curNodePointerIndex };
-  }
-  private getIndex(curNodeBucketIndex: number, curNodePointerIndex: number) {
-    if (curNodeBucketIndex === this.first) {
-      return curNodePointerIndex - this.curFirst;
-    }
-    if (curNodeBucketIndex === this.last) {
-      return this.length - (this.curLast - curNodePointerIndex) - 1;
-    }
-    return (this.bucketSize - this.first) +
-           (curNodeBucketIndex - this.first - 1) *
-           this.bucketNum + curNodePointerIndex;
   }
   clear() {
     this.first = this.last = this.curFirst = this.curLast = this.bucketNum = this.length = 0;
@@ -214,15 +201,27 @@ class Deque<T> extends SequentialContainer<T> {
   pushBack(element: T) {
     checkUndefinedParams(element);
     if (this.length) {
-      if (this.last === this.bucketNum - 1 && this.curLast === this.bucketSize - 1) {
-        this.reAllocate(this.length);
-      }
+      let changedLast = this.last;
+      let changedCurLast = this.curLast;
       if (this.curLast < this.bucketSize - 1) {
-        ++this.curLast;
+        ++changedCurLast;
       } else if (this.last < this.bucketNum - 1) {
-        ++this.last;
-        this.curLast = 0;
+        ++changedLast;
+        changedCurLast = 0;
+      } else {
+        changedLast = 0;
+        changedCurLast = 0;
       }
+      if (
+        changedLast === this.first &&
+        changedCurLast === this.curFirst
+      ) {
+        this.reAllocate(this.length);
+        this.pushBack(element);
+        return;
+      }
+      this.last = changedLast;
+      this.curLast = changedCurLast;
     }
     ++this.length;
     this.map[this.last][this.curLast] = element;
@@ -232,8 +231,11 @@ class Deque<T> extends SequentialContainer<T> {
     if (this.length !== 1) {
       if (this.curLast > 0) {
         --this.curLast;
-      } else if (this.first < this.last) {
+      } else if (this.last > 0) {
         --this.last;
+        this.curLast = this.bucketSize - 1;
+      } else {
+        this.last = this.bucketNum - 1;
         this.curLast = this.bucketSize - 1;
       }
     }
@@ -245,15 +247,27 @@ class Deque<T> extends SequentialContainer<T> {
   pushFront(element: T) {
     checkUndefinedParams(element);
     if (this.length) {
-      if (this.first === 0 && this.curFirst === 0) {
-        this.reAllocate(this.length);
-      }
+      let changedFirst = this.first;
+      let changedCurFirst = this.curFirst;
       if (this.curFirst > 0) {
-        --this.curFirst;
+        --changedCurFirst;
       } else if (this.first > 0) {
-        --this.first;
-        this.curFirst = this.bucketSize - 1;
+        --changedFirst;
+        changedCurFirst = this.bucketSize - 1;
+      } else {
+        changedFirst = this.bucketNum - 1;
+        changedCurFirst = this.bucketSize - 1;
       }
+      if (
+        changedFirst === this.last &&
+        changedCurFirst === this.curLast
+      ) {
+        this.reAllocate(this.length);
+        this.pushFront(element);
+        return;
+      }
+      this.first = changedFirst;
+      this.curFirst = changedCurFirst;
     }
     ++this.length;
     this.map[this.first][this.curFirst] = element;
@@ -266,32 +280,19 @@ class Deque<T> extends SequentialContainer<T> {
     if (this.size() !== 1) {
       if (this.curFirst < this.bucketSize - 1) {
         ++this.curFirst;
-      } else if (this.first < this.last) {
+      } else if (this.first < this.bucketNum - 1) {
         ++this.first;
+        this.curFirst = 0;
+      } else {
+        this.first = 0;
         this.curFirst = 0;
       }
     }
     if (this.length > 0) --this.length;
   }
   forEach(callback: (element: T, index: number) => void) {
-    if (!this.length) return;
-    let index = 0;
-    if (this.first === this.last) {
-      for (let i = this.curFirst; i <= this.curLast; ++i) {
-        callback(this.map[this.first][i], index++);
-      }
-      return;
-    }
-    for (let i = this.curFirst; i < this.bucketSize; ++i) {
-      callback(this.map[this.first][i], index++);
-    }
-    for (let i = this.first + 1; i < this.last; ++i) {
-      for (let j = 0; j < this.bucketSize; ++j) {
-        callback(this.map[i][j], index++);
-      }
-    }
-    for (let i = 0; i <= this.curLast; ++i) {
-      callback(this.map[this.last][i], index++);
+    for (let i = 0; i < this.length; ++i) {
+      callback(this.getElementByPos(i), i);
     }
   }
   getElementByPos(pos: number) {
@@ -366,11 +367,10 @@ class Deque<T> extends SequentialContainer<T> {
   eraseElementByValue(value: T) {
     if (!this.length) return;
     const arr: T[] = [];
-    this.forEach(element => {
-      if (element !== value) {
-        arr.push(element);
-      }
-    });
+    for (let i = 0; i < this.length; ++i) {
+      const element = this.getElementByPos(i);
+      if (element !== value) arr.push(element);
+    }
     const _length = arr.length;
     for (let i = 0; i < _length; ++i) this.setElementByPos(i, arr[i]);
     this.cut(_length - 1);
@@ -383,37 +383,17 @@ class Deque<T> extends SequentialContainer<T> {
     return iter;
   }
   find(element: T) {
-    const resIndex = (() => {
-      if (this.first === this.last) {
-        for (let i = this.curFirst; i <= this.curLast; ++i) {
-          if (this.map[this.first][i] === element) {
-            return this.getIndex(this.first, i);
-          }
-        }
-        return undefined;
+    for (let i = 0; i < this.length; ++i) {
+      if (this.getElementByPos(i) === element) {
+        return new DequeIterator(
+          i,
+          this.size,
+          this.getElementByPos,
+          this.setElementByPos
+        );
       }
-      for (let i = this.curFirst; i < this.bucketSize; ++i) {
-        if (this.map[this.first][i] === element) {
-          return this.getIndex(this.first, i);
-        }
-      }
-      for (let i = this.first + 1; i < this.last; ++i) {
-        for (let j = 0; j < this.bucketSize; ++j) {
-          if (this.map[i][j] === element) {
-            return this.getIndex(i, j);
-          }
-        }
-      }
-      for (let i = 0; i <= this.curLast; ++i) {
-        if (this.map[this.last][i] === element) {
-          return this.getIndex(this.last, i);
-        }
-      }
-      return undefined;
-    })();
-    if (resIndex === undefined) return this.end();
-    return new DequeIterator(resIndex, this.size,
-      this.getElementByPos, this.setElementByPos);
+    }
+    return this.end();
   }
   reverse() {
     let l = 0; let r = this.length - 1;
@@ -465,23 +445,8 @@ class Deque<T> extends SequentialContainer<T> {
   }
   [Symbol.iterator]() {
     return function * (this: Deque<T>) {
-      if (this.length === 0) return;
-      if (this.first === this.last) {
-        for (let i = this.curFirst; i <= this.curLast; ++i) {
-          yield this.map[this.first][i];
-        }
-        return;
-      }
-      for (let i = this.curFirst; i < this.bucketSize; ++i) {
-        yield this.map[this.first][i];
-      }
-      for (let i = this.first + 1; i < this.last; ++i) {
-        for (let j = 0; j < this.bucketSize; ++j) {
-          yield this.map[i][j];
-        }
-      }
-      for (let i = 0; i <= this.curLast; ++i) {
-        yield this.map[this.last][i];
+      for (let i = 0; i < this.length; ++i) {
+        yield this.getElementByPos(i);
       }
     }.bind(this)();
   }